Add config class to make sure the configuration is set correctly
This new class creates abstraction in the Account class. It ensures
configuration values are set (defaults are provided if they are missing
in the config file) and makes sure they are the right type. This also
means functions that use the configuration values don't have to perform
any pre-checks to make sure they are valid.