Don't try to remove flight from the account after it has checked in. Fixes #11
The problem is that the Process copies the Account into its new memory
location. When updating the Account in the main thread, the changes
aren't reflected within that Subprocess.

It isn't necessary to have this functionality at the moment, but I will
probably solve this in the future by adding a flag to the flight to
indicate if it has checked in and remove it in the Account from the main
thread if it has.