Add flight retriever to automatically retrieve flights
The parent Flight Retriever class is responsible for holding all needed
objects, such as the configuration, notification handler, and checkin
scheduler. It also allows reservations to be scheduled directly or by
JSON retrieved from an account.

The Account Flight Retriever monitors an account every 24 hours
(default, but can be adjusted). If new flights are found, it will
schedule them using the Checkin Scheduler. This class is only
instantiated when a username and password are provided.