Add note in README about using a specific version of Google Chrome in Docker
I tried to find a way to allow the user to specify a specific version of
Chrome to use. However, the repository that Google Chrome is currently
fetched from only has the latest version. Since specifying the Chrome
version is only a one-time thing, I decided not to worry about it. If
anyone else figures out how to do it though, a PR is welcome.