Add Docker image to public repository
This commit adds a workflow to automatically build the Docker image
whenever a release is published (with tags ':latest' and
':vX.X.X'). Additionally, a Docker image is published every push in the
develop branch with the ':develop' tag.

Update the documentation to point users to download from the public
repository instead of building it themselves.

The one problem with the Docker image being downloaded instead of built
is that the user doesn't know which version Google Chrome is. To handle
this, a `docker-entrypoint.sh` script was added to automatically
retrieve the currently downloaded version of Chrome and set an environment
variable for the script to read.