Ensure webdriver requests are empty before setting its scope
Fixes #74.

Sometimes, Chrome can make requests as it is being initialized. These
requests are not filtered by the `driver.scopes` variable as it isn't
set yet. Therefore, the script would fail because it would select the
wrong requests.

Now the requests are deleted after initialization to ensure only
requests we want are captured.