Handle malformed URLs when making requests
Sometimes, the Southwest API could return part of a URL that has a
leading slash. When creating the full URL, a double slash would appear,
making the request fail.

Here the double slash is first replaced and then a leading slashed is
removed. This will ensure that the URL will not have any double slashes
(except for the https:// protocol).