Handle fares no longer existing for a flight
When a flight is sold out of the current fare type the flight was
purchased under, a fare difference is no longer returned from Southwest.
Therefore, set the difference to 0 and return that.

There is no reason to raise an error here (and catch it later) because the
fare check never really failed. The price difference is 0 USD if it doesn't
exist as the user can't change the flight using the same fare.