Don't attempt to check fares if the change link is not available
Reservations can have a message that pops up when the Change button is
clicked that says "Flight cannot be changed online. Contact ... for more
information" and the change link is null.

Here, a check is added to ensure the change link is available before
proceeding further in the fare check.