Fix fares not being compared for the correct flight in some cases
Previously, the matching flight was only retrieved based on the
having the same departure time as the flight being checked for fares.
This created problems if two different flights had the same departure
time. Now, both the departure time and arrival time are checked to
ensure the correct flight is selected.