Improve Ctrl-C handling in Windows processes
Permission errors when killing a process are now captured as well. I
believe this happens because the process no longer exists on Windows. I
still get one error (OSError: [WinError 10038] An operation was attempted
on something that is not a socket), but I haven't been able to figure
out how to fix that. It is pretty irrelevant though as it still prints
the output well (nor will this part of the code virtually ever be
reached)