Fix check-ins for same-day flights
Should fix #135.

A same-day flight is a flight that is within 24 hours of the first
flight on a round-trip reservation.

Since this scripts attempts to check in five seconds early, the check-in
for a same-day flight (the second flight of a round trip) would succeed
right away due to the first flight already being checked in (but the
second flight would not have checked in yet).

Now, if it is a same-day flight, the script will attempt to check in
until both flights are sucessfully checked in.