Add integration tests
These automated tests run many parts of the script together. These tests
will help ensure the program works correctly together, and not just in
individual units.

The integration tests do not test every possible situation in the
script, but the important situations are tested to ensure everything
works as expected.