Make both a GET and POST request for check-ins every attempt
Previously, someone attempted a same-day flight check-in with just
repeatedly doing the POST request. However, the check-in never went
through for them. I believe it is because the check-in info from the GET
request changed when the new flight became available. Therefore, this
commit changes the check-in process to make both the GET and POST
request repeatedly to hopefully solve this issue.