Install the latest minor chromedriver version for a milestone
Fixes #180

Directly matching the browser version appeared problematic when the
latest browser version did not have a matching driver version. Now, the
latest version is installed, while still ensuring there is the same
major version (e.g. browser v117.* and driver v117.*).

To not download the chromedriver in Docker images (the driver download
is very finicky with Alpine Linux), an environment variable in the
Docker image is set and read in the script to keep the downloaded
chromedriver, as we want that to be set up in the image, not the script.