Handle flight departure time changes correctly
Fixes #199.

This is actually a regression that was implemented in v6.1 when I
changed to differentiate flights by flight number. The issue is that
a flight can still have the same flight number but not the same
departure time (if it was rescheduled). Now, both the flight number and
departure time are checked to see whether the flight is equal to one
already scheduled for check in.