Fix check-ins for the second leg of same-day flights (#241)
* Fix check-ins for same-day flights

Should fix #135.

A same-day flight is a flight that is within 24 hours of the first
flight on a round-trip reservation.

Since this scripts attempts to check in five seconds early, the check-in
for a same-day flight (the second flight of a round trip) would succeed
right away due to the first flight already being checked in (but the
second flight would not have checked in yet).

Now, if it is a same-day flight, the script will attempt to check in
until both flights are sucessfully checked in.

* Add an integration test for same-day flight check-ins

* Make both a GET and POST request for check-ins every attempt

Previously, someone attempted a same-day flight check-in with just
repeatedly doing the POST request. However, the check-in never went
through for them. I believe it is because the check-in info from the GET
request changed when the new flight became available. Therefore, this
commit changes the check-in process to make both the GET and POST
request repeatedly to hopefully solve this issue.

* Keep all flights on the reservation when checking in

Keeping only the last flight had issues with a connecting flight
boarding position not showing up (it only showed the last one, not all).

* Add more robust test validation for same day flights