Improve departed flight checking
In #257, the user was getting a departure status of None for a flight
that was from last week instead of "DEPARTED". The reason is unknown but
it could be because it departed so long ago (I have a flight I can test
on but will need to wait a week).

Now, a flight will only be scheduled if its departure time is in the
future