Randomly sleep by default when making a request
Almost all requests need to have random sleep between failed attempts,
so the code is less verbose by having it default to random sleep.
Additionally, if requests are added in the future it is likely that they
will need to also use random sleep.

The only request not using random sleep is the POST request when
checking in as that needs to be done as fast as possible (and doesn't
seem to have the same bot detection that the other endpoints do).