Always write logs under the "logs" directory in the root of the project
Before, the script would create the logs directory and write logs in the
directory the script was executed from. Now, the logs are always written
under the 'logs' directory in the project's root directory.

The script still can't be run from anywhere because SeleniumBase writes
a 'downloaded_files' folder to the directory the script is executed
from but this change is still an improvement.