Use caching to speed up fare checks
Because a flight's reservation info is retrieved in the checkin scheduler
right before the fare checker retrieves the reservation info, this info
can be cached to remove the need for the fare checker to request this
info. This improves the speed of the fare checker and reduces strain
(and chance of detection) on Southwest's API due to the 33% decrease in
fare checker requests per flight.