Add a timeout for waiting for headers or a login
This prevents the webdriver waiting in an infinite loop for the headers
request, login response, or trips response to come through.

If a timeout occurs while retrieving reservations, the retrieval will be
skipped and continue again on the next retrieval interval. If the timeout
occurs while refreshing headers for the check-in, the check-in will
continue but will probably fail (in later commits, a notification will
be sent when this happens).