Add notifications for notices
Notices are warnings that are non-critical. These include driver
timeouts during reservation retrieval (a timeout before a check-in will
be sent as an error) and a Too Many Requests error during a login.

This is a breaking change for the notification_level config value. If it
is set manually, it will need to be adjusted to the new scale in
CONFIGURATION.md.