Add a timeout when waiting for headers and logins to go through and a new notice notification level (#281)
* Add a timeout for waiting for headers or a login

This prevents the webdriver waiting in an infinite loop for the headers
request, login response, or trips response to come through.

If a timeout occurs while retrieving reservations, the retrieval will be
skipped and continue again on the next retrieval interval. If the timeout
occurs while refreshing headers for the check-in, the check-in will
continue but will probably fail (in later commits, a notification will
be sent when this happens).

* Add notifications for notices

Notices are warnings that are non-critical. These include driver
timeouts during reservation retrieval (a timeout before a check-in will
be sent as an error) and a Too Many Requests error during a login.

This is a breaking change for the notification_level config value. If it
is set manually, it will need to be adjusted to the new scale in
CONFIGURATION.md.

* Prefix error notifications with 'Error:'

This follows along with the notice notifications and makes them easier
to identify right away when reading the message (although, they are
still easy to identify because they all contain some version of 'fail')

* Update changelog to include information about the new notification
levels