Remove any non-number characters from flight numbers
Fixes #288.

Southwest now includes a prefix to the start of their flight numbers.
Instead of '100' it is now 'WN100'. I'm not sure if the only prefix is
WN, so I opted to remove all non-number characters as the fare checking
still requires only numbers. Eventually, this will probably change and
these changes can be reverted.