Simplify removing 'WN' prefix for each flight number
WN is Southwest's IATA code, so I think it is safe to assume that only
WN will be in the flight number. Therefore, simplify the code to only
remove the WN prefix.

If Southwest changes this again in the future, it will still be good for
the script to break so we know exactly what changed (and it would be an
easy fix).