Allow for setting notification levels per URL
The notification configuration has been improved to allow for setting
notification levels per URL. This change is not backwards-compatible.

The new configuration format is as follows:
"notifications": [{"url": "sms://text", "level": 1, "24_hour": true}]

The "url" field is required and must be a valid URL. The "level" field
is optional and defaults to INFO. The "24_hour" field is optional and
defaults to false.

To fully finish this feature, three things still need to be done:
- Adapt the NotificationHandler class to use this new configuration
- Change the way environment variables are parsed to account for this
  change
- Update the documentation to reflect this new feature