Check in using multiple threads, each starting one second after the other
This is an attempt at fixing #325 and improving the speed of check-ins.
During check-in, what usually happens is that a bad request (400) will
go through on the first attempt. This will then take ~3 seconds to
finish before the next request is attempted.

This change will allow multiple threads to check in at the same time,
each starting one second after the other. This will reduce the time
between check-in requests and hopefully reduce the total time it takes
to check in.